Preparatory refactoring to add grpc-fallback support
